Remove commit from cross-repo ostree updates
When creating a cross-repo ostree update, do not create a new commit
in the base repo. Simply generate a diff to the existing to_sha.

This avoids the problem of upgrade archives having different SHAs to
the originally-compiled WIC images, but at the expense of the upgrade
ostree no longer having a linear history, if the new repo was not based
on the old one.

In this version, the to_sha commit is pulled into the base repo, but no
refs are modified.

Subsequent work will determine how to manage the base repo - you might
for example want to reset the branch in the base repo to that from the
upgrade repo, and save the old branch as <machine>_<date>.